python/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
python-dotenv==1.0.1
livekit-api==0.8.2
livekit

